+++ ml/src/main/java/kr/gaion/armoredVehicle/database/model/SensorGearboxLife.java
@@ -27,7 +27,7 @@ public class SensorGearboxLife {
     private Date operateDateTime;
 
     @Column(name = "AI_Trip")
-    private Integer aiTrip;
+    private Double aiTrip;
 
     @Column(name = "AI_Trip_ALGO")
     private String aiTripAlgo;
